fix: resolve CI failures for runtime path layout and backend size
Address Windows test flakes from exclusive-create placeholder files, update functional API log path assertions for YYYY/MM/DD layout, and split oversized host_children tests to satisfy backend-size limits.
